Address review: extract ext::lift, tighten qi fallback and claims
Move `Liftable`, `LiftRequest` and `MultiLift` out of
`resolution_homomorphism` into a new `ext::lift` module, and collapse the
three identical `StepPrep`/`HomotopyPrep`/`SecondaryHomotopyPrep` enums
into one generic `LiftPrep<P>` there.

`apply_quasi_inverse` no longer falls back to recompute unconditionally:
a save store that should hold the quasi-inverse but doesn't is an
incomplete store, and reports failure as it did before
recompute-on-demand existed. Recompute stays for the forced, never
persisted, and legitimately-unsaved-top-of-range cases.

`MultiLift::new` now asserts every liftable shares the target, via a new
`Liftable::target_addr`; batching against a foreign target would solve
the steps with the wrong quasi-inverse.

The rest is documentation the review found overclaiming or misplaced:
`RecomputeReader`'s peak-memory claim now accounts for the serialized
block, `MultiLift::extend_through_stem` says it bounds output rather than
source bidegrees, `extend_all_secondary_products` states its
positive-filtration restriction, and the perf figures and experiment-log
prose come out of doc comments.
