+++ ext/src/ext_algebra/mod.rs
@@ -31,7 +31,8 @@ use sseq::coordinates::{Bidegree, BidegreeElement, BidegreeGenerator};
 pub use self::secondary::{SecondaryExtAlgebra, SecondaryProduct};
 use crate::{
     chain_complex::{AugmentedChainComplex, FreeChainComplex},
-    resolution_homomorphism::{Liftable, MultiLift, ResolutionHomomorphism},
+    lift::{Liftable, MultiLift},
+    resolution_homomorphism::ResolutionHomomorphism,
     utils::{QueryModuleResolution, get_unit},
 };
 
@@ -201,9 +202,7 @@ where
     ///
     /// All the product maps share the unit resolution as target, so lifting them together means the
     /// unit's quasi-inverse at each output bidegree is solved once and reused across every product,
-    /// rather than once per product. With quasi-inverses recomputed on demand this keeps the cost of
-    /// computing all products at ~1x the on-disk-quasi-inverse baseline instead of scaling with the
-    /// number of generators.
+    /// rather than once per product.
     pub fn extend_all_products(&self)
     where
         CC: Sync + 'static,
